Cast pointer-offset indices to int64 in batched V1 and PQ kernels
Extends the int64 fix from #1 to the two documented kernels it did not
touch: the batched V1 MaxSim kernel (flash_maxsim_kernel.py) and the
batched PQ kernel (flash_pqsim_kernel.py), plus their single-query
variants for consistency.

Both had the same latent signed int32 overflow in pointer arithmetic
(batch_idx * Nd * d and batch_idx * Nd * M), causing CUDA illegal memory
access on large batches before GPU memory is exhausted. Casting the
program ids and index vectors to int64 promotes the offset computation
and removes the overflow.

Validated on H100 (Nq=44, Nd=747, d=320, fp16): scores match the PyTorch
reference (max abs err 4.9e-4), and both kernels run correctly past their
overflow thresholds (MaxSim ~8984 docs, PQ ~71871 docs) up to 32768 and
100000 documents respectively.

+++ src/flash_maxsim_kernel.py
@@ -54,7 +54,8 @@ def _flash_maxsim_fwd_kernel(
 
     # Iterate over document token tiles
     for d_start in range(0, Nd, BLOCK_Nd):
-        d_indices = d_start + tl.arange(0, BLOCK_Nd)  # (BLOCK_Nd,)
+        # int64 so d_indices * d cannot overflow signed int32 for large Nd
+        d_indices = (d_start + tl.arange(0, BLOCK_Nd)).to(tl.int64)  # (BLOCK_Nd,)
         d_mask = d_indices < Nd
 
         # Compute dot products between this query token and document token tile
@@ -106,21 +107,21 @@ def _flash_maxsim_batched_kernel(
     Batched version: each program handles one (batch, query_token) pair.
     Final reduction (sum over query tokens) done in a separate small kernel.
     """
-    batch_idx = tl.program_id(1)
+    batch_idx = tl.program_id(1).to(tl.int64)
     q_idx = tl.program_id(0)
 
     if q_idx >= Nq:
         return
 
-    # Offset into batch
+    # Offset into batch (int64 to avoid signed int32 overflow on large batches)
     Q_batch = Q_ptr + batch_idx * Nq * d
     D_batch = D_ptr + batch_idx * Nd * d
 
     running_max = tl.full([], value=float('-inf'), dtype=tl.float32)
-    q_offsets = tl.arange(0, BLOCK_d)
+    q_offsets = tl.arange(0, BLOCK_d).to(tl.int64)
 
     for d_start in range(0, Nd, BLOCK_Nd):
-        d_indices = d_start + tl.arange(0, BLOCK_Nd)
+        d_indices = (d_start + tl.arange(0, BLOCK_Nd)).to(tl.int64)
         d_mask = d_indices < Nd
         dots = tl.zeros([BLOCK_Nd], dtype=tl.float32)
